+++ parser_library/include/protocol.h
@@ -401,10 +401,18 @@ using variables_t = sequence<variable, const debugging::variable_store*>;
 
 struct breakpoint
 {
-    breakpoint(size_t line)
+    explicit breakpoint(size_t line)
         : line(line)
     {}
     size_t line;
+};
+
+struct function_breakpoint
+{
+    explicit function_breakpoint(sequence<char> name)
+        : name(name)
+    {}
+    sequence<char> name;
 };
 
 struct output_line