+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Watch support in the Macro tracer
 - Emit MNOTE and PUNCH arguments to the debug console
 - Make MNOTE and PUNCH outputs available from VSCode
+- Function breakpoint support in the Macro tracer
 
 #### Fixed
 - Unknown requests were dropped without a proper response